Moved src/Table/write_ipac_table directory to src/ipac_table_writer and split off new ipac_table_writer namespace from Table.
